requirements.txt
discord.py
urlextract
pytorch